chore: rename with kebab convention
